refactor(agent-core): 拆分 runtime parity smoke 用例
